+++ java/core/src/main/java/io/v6d/core/client/IPCClient.java
@@ -93,7 +93,6 @@ public class IPCClient extends Client {
         this.doWrite(root);
         val reply = new CreateDataReply();
         reply.get(this.doReadJson());
-        System.out.println("######## " + reply.getId());
         if (meta.isIncomplete()) {
             return getMetaData(reply.getId());
         } else {
